- Updated int[] to byte[] - Changes i + 1 index to i index

+++ game_loop.h
@@ -4,8 +4,8 @@
 #include "74hc595.h"
 #include "config.h"
 
-extern int obj_list[OBJ_NBR];
+extern byte obj_list[OBJ_NBR];
 extern int _difficulty;
 
-void  shuffleList(int arrayToShuffle[], int size);
+void  shuffleList(byte arrayToShuffle[], int size);
 void  game_loop();
